requirements.txt
Django==2.0.6
django-ckeditor==5.6.1
django-js-asset==1.1.0
mysqlclient==1.3.13
Pillow==5.2.0
python-decouple==3.1
pytz==2018.5
whitenoise==3.3.1
